User Tracking v1.5.2 - The step back to 1.4.2
This "update" was pushed to ZC as a version upgrade to address the install
sql which could be finicky and hadn't been previously well enough written
to support all installations. It needed work, there was an example in the
forum, but it had been decided by corseter to upload an old version with
a new install sql file.  There were several things that were undone in
this update including additional data sanitization to prevent system
attack to be the most troubling.