Prevent install auto_loader from executing on catalog side.
Capturing previous problem with installer identified with
another plugin that the installation should be
prevented if the auto_loader file(s) are copied to the wrong side
of the store. At least this way it does not
repetitively execute, could log a message; however,
have not gone so far yet.