Delete HTML commented code from PHP section
This fixes #1.  

I decided to delete the line as it is something presented early on in the generation of the tables and does not appear to offer any significant benefit to the page source.
The history is that in transitioning from the standard html table structure to the ZC base table style/method to possibly support improved capability/operability with ZC 1.6 I considered that this style of box would be an improvement for the output.  Some further internal alignment/structure/readability could be afforded, but as a result of attempting to maintain the source code possibly displayed with the commented out portions in addition/parallel with the table format, this line (390) did not have the correct open/close statement maintained with the update.