Installer revision.
Added some refinements to the installer inlcuding
assignment of the variable to control whether to
perform version checking specific to the plugin when
ZC also allows version checking.

This means that if ZC is not performing version checks
then the plugin also will not perform a version check, if
ZC performs version checking, then the plugins check can
be turned off or kept on.