Add instructions that would apply to ZC 1.5.1
In ZC 1.5.2, a change was made to admin/includes/classes/table_block.php
that made it possible to finally show text in boxes.  This
plugin has moved to using such a box for everything and as
a result, has caused the "disappearance" of the content
when using an older version of the ZC table/box code.