Correct usage of array variable.
$user_tracking was not previously declared as an array
so when evaluated for an array condition would send a
user warning notice to be logged. This corrects the usage
so that $user_tracking is declared as an array but where
code was previously not to be executed because the
variable was not declared an array, that has been changed
to validate that the array is not empty (meaning if
it has been set to more than just the assignment of
array() as one characteristic).